chore: separate NotFoundComponent
