+++ src/components/NotFound/index.tsx
@@ -1,0 +1,1 @@
+export { default as NotFoundComponent } from './NotFoundComponent';
